Reorganized search page files
