+++ frontend/components/app.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Modal from './modal/modal'
 import NavBarContainer from './nav_bar/nav_bar_container'
 import SplashContainer from './splash_page/splash_container'
-import SearchPageContainer from './search/search_page_container'
+import SearchPageContainer from './search_page/search_page_container'
 import RestaurantShowContainer from './restaurant_show/restaurant_show_container'
 import ProfileContainer from './profile/profile_container'
 import NavFooter from './nav_footer/nav_footer'
