+++ frontend/components/search_page/search_page.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import SearchFormContainer from './search_form_container'
+import SearchFormContainer from '../search_form/search_form_container'
 import RestaurantSearch from '../restaurant_search/restaurant_search'
 import FilterForm from './filter_form'
